pyproject.toml
[project]
name = "mora"
version = "0.1.0"
description = "Source code of paper MoRA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "filelock==3.10",
    "hydra-core>=1.3.2",
    "jsonlines>=4.0.0",
    "loguru>=0.7.3",
    "nni==2.7",
    "pyarrow>=19.0.1",
    "pyrootutils>=1.0.4",
    "torch>=2.6.0",
    "torchmetrics>=1.6.2",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "typeguard<3",
]
